Added JWTAuthentication through the simpleJWT third party tool. The settings were updated to use it by default, however it was removed in favor of adding it later if needed.
In the serializers.py file, the opponent_name was added as a CharField for responses. Initially it was a ReadOnlyField which would replace the opponent tag, however it was changed back and added separately into the fields for testing purposes.

Permission for anyone to make an account was added. Line 21: authentication_classes = [] is set to disable any kind of default authentication required in case it is setup inside of settings.py. Without it, default authentication would have taken over and a user would not be able to register without already having an account which is paradoxical.

Views were added for Listing games that require authentication through JWT. Additionally SessionAuthentication is being considered in order to access the tools through a browser instead of just curl/httpie requests.

JoinGame was added as an APIView. The patch method is used in order for the opponent to join a match and afterwards play. Operations regarding opening games, listing them and similiar are being considered to be bundled up inside a ViewSet and routed later using a router.