Added pagination, spectacular and JWT authentication app wide to the settings
Added a filter class using the django_filters module which allows for filtering based on opponents, creators, winners and game_states

The model has been slightly altered. The related_names were changed for creator and opponent. Winner was changed from boolean to ForeignKey in order to make lookups easier.

Added a MoveSerializer for drf-spectacular swaggerUI

The UserSerializer was changed in order to acommodate the project requirements. Upon getting the data from Users, number of games and the percentage of games won returns together with the user upon request. Implemented by overriding the get_xx methods.

URLs for game specific actions were added like:
 - checking the board
 - joining a game
 - making a move

Additional urls for user profiles and drf-spectacular specific processes added

utils.py was added to hold the tictactoe game logic class.

Views were altered by removing the JWTAuthentication off of classes. Since it is globally set, there is no need to hold them there. UserProfile was added as a RetrieveApiView with a lookup_field based on the username. Filters were added to the GamesList ListAPIView. An APIView for joining a game was added. A function-based view was made for making a move. It hold custom logic hence why it was added like that.

Due to changes in the model, another migration had to be made.